Separate Zyte-specific code into spidermon.contrib.zyte
