+++ spidermon/contrib/zyte/monitors/mixins/__init__.py
@@ -1,0 +1,3 @@
+from .job import JobMonitorMixin
+
+__all__ = ["JobMonitorMixin"]
