+++ spidermon/contrib/zyte/statscollectors/__init__.py
@@ -1,0 +1,3 @@
+from .sc_collections import ScrapyCloudCollectionsStatsHistoryCollector
+
+__all__ = ["ScrapyCloudCollectionsStatsHistoryCollector"]
