+++ tests/contrib/scrapy/monitors/test_zyte_deprecated.py
@@ -1,0 +1,13 @@
+import pytest
+
+pytest.importorskip("scrapy")
+
+from spidermon.contrib.scrapy import monitors
+
+
+def test_deprecated_attribute():
+    with pytest.warns(
+        DeprecationWarning,
+        match="spidermon.contrib.scrapy.monitors.ZyteJobsComparisonMonitor",
+    ):
+        assert monitors.ZyteJobsComparisonMonitor is not None
